# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hotel CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hotels:
    get:
      operationId: listHotels
      tags: [hotels]
      summary: List all hotels
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"
    post:
      operationId: createHotel
      tags: [hotels]
      summary: Create a hotel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHotelRequest"
      responses:
        "201":
          description: Hotel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
  /hotels/{id}:
    get:
      operationId: getHotel
      tags: [hotels]
      summary: Get a hotel by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hotel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
    put:
      operationId: updateHotel
      tags: [hotels]
      summary: Update a hotel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHotelRequest"
      responses:
        "200":
          description: Hotel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
    delete:
      operationId: deleteHotel
      tags: [hotels]
      summary: Delete a hotel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Hotel deleted
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: hotelId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking (Owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Booking deleted
  /bookings/{id}/payment:
    patch:
      operationId: updatePayment
      tags: [bookings]
      summary: Update receipt/payment for a booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentRequest"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard stats
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/checkins:
    get:
      operationId: getTodayCheckins
      tags: [dashboard]
      summary: Get today's check-ins
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Today's check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /dashboard/checkouts:
    get:
      operationId: getTodayCheckouts
      tags: [dashboard]
      summary: Get today's check-outs
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Today's check-outs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        totalRooms:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - totalRooms
        - createdAt
    CreateHotelRequest:
      type: object
      properties:
        name:
          type: string
        totalRooms:
          type: integer
      required:
        - name
        - totalRooms
    Agency:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        hotelId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - hotelId
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        roomNumber:
          type: string
          nullable: true
        roomType:
          type: string
          nullable: true
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        roomRent:
          type: number
        addOns:
          type: number
        totalCost:
          type: number
        receipt:
          type: number
        balance:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [confirmed, checked_in, checked_out, cancelled]
        hotelId:
          type: integer
        agencyId:
          type: integer
          nullable: true
        agency:
          $ref: "#/components/schemas/Agency"
          nullable: true
        hotel:
          $ref: "#/components/schemas/Hotel"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - guestName
        - checkIn
        - checkOut
        - roomRent
        - addOns
        - totalCost
        - receipt
        - balance
        - status
        - hotelId
        - createdAt
        - updatedAt
    CreateBookingRequest:
      type: object
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        roomNumber:
          type: string
          nullable: true
        roomType:
          type: string
          nullable: true
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        roomRent:
          type: number
        addOns:
          type: number
        receipt:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [confirmed, checked_in, checked_out, cancelled]
        hotelId:
          type: integer
        agencyId:
          type: integer
          nullable: true
      required:
        - guestName
        - checkIn
        - checkOut
        - roomRent
        - addOns
        - receipt
        - status
        - hotelId
    UpdateBookingRequest:
      type: object
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        roomNumber:
          type: string
          nullable: true
        roomType:
          type: string
          nullable: true
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        roomRent:
          type: number
        addOns:
          type: number
        receipt:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [confirmed, checked_in, checked_out, cancelled]
        agencyId:
          type: integer
          nullable: true
    UpdatePaymentRequest:
      type: object
      properties:
        receipt:
          type: number
      required:
        - receipt
    DashboardStats:
      type: object
      properties:
        todayCheckins:
          type: integer
        todayCheckouts:
          type: integer
        totalBookings:
          type: integer
        occupiedRooms:
          type: integer
        totalRooms:
          type: integer
        occupancyPercentage:
          type: number
        monthlyRevenue:
          type: number
      required:
        - todayCheckins
        - todayCheckouts
        - totalBookings
        - occupiedRooms
        - totalRooms
        - occupancyPercentage
        - monthlyRevenue
...