# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { Agency } from "./agency";
import type { BookingStatus } from "./bookingStatus";
import type { Hotel } from "./hotel";

export interface Booking {
  id: number;
  guestName: string;
  guestEmail?: string | null;
  guestPhone?: string | null;
  roomNumber?: string | null;
  roomType?: string | null;
  checkIn: Date;
  checkOut: Date;
  roomRent: number;
  addOns: number;
  totalCost: number;
  receipt: number;
  balance: number;
  notes?: string | null;
  status: BookingStatus;
  hotelId: number;
  agencyId?: number | null;
  agency?: Agency | null;
  hotel?: Hotel | null;
  createdAt: Date;
  updatedAt: Date;
}