# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingRequestStatus } from "./createBookingRequestStatus";

export interface CreateBookingRequest {
  guestName: string;
  guestEmail?: string | null;
  guestPhone?: string | null;
  roomNumber?: string | null;
  roomType?: string | null;
  checkIn: Date;
  checkOut: Date;
  roomRent: number;
  addOns: number;
  receipt: number;
  notes?: string | null;
  status: CreateBookingRequestStatus;
  hotelId: number;
  agencyId?: number | null;
}